+++ CHANGELOG.md
@@ -4,10 +4,23 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [8.0.0] - 2021-06-06
+
+### Added:
+- Recipe interface that can be overrided
+- `preAPIHook` and `onHandleEvent` functions
+
+### Changed:
+- `sessionScope` is a now a string
+
+### Removed:
+- Backward compatibility with cross domain localstorage
+- Removes `setAuth0API`, `getAuth0API` and `getRefreshURLDomain` functions.
+- Removed `refreshAPICustomHeaders` and `signoutAPICustomHeaders` from config. Use `preAPIHook` instead.
+
 ## [7.2.1] - 2021-06-11
 ### Fixes:
 - Fixes issue https://github.com/supertokens/supertokens-node/issues/134
-
 
 ## [7.2.0] - 2021-06-05
 ### Added: